feat(validation): Min/max validations for OAS3
